Publish SR output in `{meta.id}/spaceranger/outs`
Publish Space Ranger count output in `{meta.id}/spaceranger/outs`
instead of just `{meta.id}/outs`, which adheres better to the
"results/{software}" strategy for other modules.

This was hard-coded in the Space Ranger module as `{meta.id}/outs`. This
commit achieves the change by modifying the module code to instead be
just `outs`, so that the exact publishDir may be specified in
`modules.config` as `${params.outdir}/{meta.id}/spaceranger` instead.
This did not work before, as modifying `modules.config` could just add
paths on top of the already hard-coded `${params.outdir}/{meta.id}/outs`
module publishDir.

A PR with these changes will be submitted to the modules repo.