Remove unintuitive `st_preprocess_fig_size` param
Remove the unintuitive `st_preprocess_fig_size` parameter. This was only
used one occasion in the pipeline, where it only affected a single
report figure - all other figures hard hard-coded sizes. This brings all
figures in line in that all sizes are hard-coded.

This should possibly be changed in the future, but giving users the
ability to change figure sizes on a pipeline-level would be difficult,
since it may be that not all figures can look good with the same size
across the board.