Change name to `nf-core/spatialvi`
As per discussions in the nf-core community, this commit changes the
name of the pipeline from `nf-core/spatialtranscriptomics` to
`nf-core/spatialvi`. The old name is descriptive in regards to what it
does, but not specific enough, as it analyses 10X Visium data
exclusively. It was determined that spatial nf-core pipelines should be
named according to the `spatialXX` pattern; another pipeline named
`nf-core/spatialxe` for Xenium data already exists.