Fix test profile resource limits
The `spaceranger/count` module stalled during GHA CI tests due to some
issue with the new template resource specifications for the test
profile. The new template specifies 15 GB of RAM, but the GHA runner
only had 13.6 available, which somehow lead to a stall of the task until
the time ran out. Setting a slightly lower memory requirement fixes
this.