Use older version of Space Ranger Docker image
Space Ranger requires the CPU to work with AVX instructions since
version `3.1.3`, which makes it and subsequent versions incompatible
with newer Apple computers (which uses Apple's own chips). We want to
keep compatibility as long as that is feasible without missing out on
features or crucial bugfixes, so while we keep to an earlier version for
now we will most likely have to forego Apple compatibility sometime in
the future.